+++ apps/web/app/(game)/(village-slug)/(players)/(...player-slug)/page.tsx
@@ -175,13 +175,15 @@ const PlayerPage = ({ params }: Route.ComponentProps) => {
                     </TableCell>
                     <TableCell>
                       <Text>
-                        <Resources
-                          className="justify-center"
-                          iconClassName="size-4"
-                          resources={parseResourcesFromRFC(
-                            resourceFieldComposition,
-                          )}
-                        />
+                        <span className="inline-flex gap-2">
+                          <Resources
+                            className="justify-center"
+                            iconClassName="size-4"
+                            resources={parseResourcesFromRFC(
+                              resourceFieldComposition,
+                            )}
+                          />
+                        </span>
                       </Text>
                     </TableCell>
                   </TableRow>